requirements.txt
apache-airflow
openai
pandas
matplotlib
sqlalchemy
prometheus_client
fastapi
uvicorn
pydantic
stateflow
